% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoStardust.R
\name{autoStardust}
\alias{autoStardust}
\title{Automatic Stardust}
\usage{
autoStardust(countMatrix, spotPositions, pcaDimensions = 10, res = 0.8)
}
\arguments{
\item{countMatrix, }{dataframe with spots id as columns and genes id as rows}

\item{spotPositions, }{dataframe with x and y spot coordinates as columns and spots id as rows}

\item{pcaDimensions, }{number of PCA dimension to include in the distance computation}

\item{res, }{double resolution for Louvain algorithm}
}
\value{
Seurat object with cluster identifications for each spot under active.ident slot.
}
\description{
This function calls Seurat clustering feature, passing a custom distance matrix that integrates both
transcriptional similarity and spatial information.
}
\author{
Eva Viesi, eva [dot] viesi [at] univr [dot] it, University of Verona

Simone Avesani, simone [dot] avesani [at] univr [dot] it, University of Verona

Giovanni Motterle, giovanni [dot] motterle [at] univr [dot] it, University of Verona
}
