% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouseKidney-data.R
\docType{data}
\name{MouseKidney}
\alias{MouseKidney}
\title{Mouse Kidney gene expression matrix}
\format{
An object of class \code{"dataframe"}.
}
\source{
\href{https://www.10xgenomics.com/resources/datasets/}{10X website, dataset archive}
}
\usage{
data(MouseKidney)
}
\description{
Each column in this dataset represents the transcriptional
profile of one spot in a dataset of Spatial Transcriptomics
generated from a tissue section of a mouse kidney
}
\keyword{datasets}
