% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouseKidneyCoord-data.R
\docType{data}
\name{MouseKidneyCoord}
\alias{MouseKidneyCoord}
\title{Mouse Kidney spots coordinates}
\format{
An object of class \code{"dataframe"}.
}
\source{
\href{https://www.10xgenomics.com/resources/datasets/}{10X website, dataset archive}
}
\usage{
data(MouseKidneyCoord)
}
\description{
Each row in this dataset represents the physical
position of one spot in a dataset of Spatial Transcriptomics
(coming from 10X Visium assey) generated from a tissue section 
of a mouse kidney
}
\keyword{datasets}
