% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StardustOnSeurat.R
\name{StardustOnSeurat}
\alias{StardustOnSeurat}
\title{Stardust on Seurat}
\usage{
StardustOnSeurat(
  countMatrix,
  spotPositions,
  spaceWeight = 1,
  pcaDimensions = 5
)
}
\arguments{
\item{countMatrix, }{dataframe with spots id as columns and genes id as rows.}

\item{spotPositions, }{dataframe with x and y spot coordinates as columns and spots id as rows.}

\item{spaceWeight, }{double in [0,1]. Weight for the linear transformation of spot distance.
1 means spot distance weight as much as profile distance, 0 means spot distance doesn't contribute at all in the overall 
distance measure.}

\item{pcaDimensions, }{number of PCA dimension to include in the distance computation.}
}
\value{
Seurat object with cluster identifications for each spot under active.ident slot.
}
\description{
This function calls Seurat clustering feature, passing a custom distance matrix that integrate both
transcriptional similarity and spatial information
}
\author{
Giovanni Motterle, giovanni [dot] motterle [at] univr [dot] it, University of Verona
}
