% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightStardust.R
\name{weightStardust}
\alias{weightStardust}
\title{Weighted Stardust}
\usage{
weightStardust(
  countMatrix,
  spotPositions,
  spaceWeight = 1,
  pcaDimensions = 10,
  res = 0.8
)
}
\arguments{
\item{countMatrix, }{dataframe with spots id as columns and genes id as rows.}

\item{spotPositions, }{dataframe with x and y spot coordinates as columns and spots id as rows.}

\item{spaceWeight, }{double in [0,1]. Weight for the linear transformation of spot distance.
1 means spot distance weight as much as profile distance, 0 means spot distance doesn't contribute at all in the overall 
distance measure.}

\item{pcaDimensions, }{number of PCA dimension to include in the distance computation.}

\item{res, }{double resolution for Louvain algorithm.}
}
\value{
Seurat object with cluster identifications for each spot under active.ident slot.
}
\description{
This function calls Seurat clustering feature, passing a custom distance matrix that integrates both
transcriptional similarity and spatial information through a space weight parameter.
}
\author{
Eva Viesi, eva [dot] viesi [at] univr [dot] it, University of Verona

Simone Avesani, simone [dot] avesani [at] univr [dot] it, University of Verona

Giovanni Motterle, giovanni [dot] motterle [at] univr [dot] it, University of Verona
}
